Clean up some markdown and fix broken links to images in Section 4.
Initial clean up of Section 4 to add sub-section numbering and titles, remove some artifacts from the pandoc conversion, and fix the broken image links.  Note that many artifacts remain, and another scrub of Section 4 is warranted.  Also noting that the amount and size of the images in Section 4 creates a fairly large and slow-responding file (at least when knit in Rstudio).  Replacing equation figures with math-mode text may help, as will reducing the image sizes.