Replace images of equations with LaTeX in Sections 4.1 through 4.4  and various cleanup of pandoc conversion artifacts
The remaining subsections of Section 4 still need to be updated in a similar way.  Images of equations that are no longer being used have been removed from the docs/media directory, but have not been removed from the handbookdown/media directory.  Also, references to these images at the bottom of the .Rmd file have not been removed yet.