Replace figures in Sections 4.5 and 4.6 with text and LaTeX equations
Note that I took some minor liberties with the presentation of these equations and figures.  For instance, in Section 4.5, the majority of the "text" was actually part of a figure.  I cropped the figure to only the non-text part of the image, then added the plain text back in the .Rmd file itself.  A few minor things were changed; I couldn't find a way to generate a "therefore" 3-triangle symbol, so the word "therefore" was used instead.  Additionally, I couldn't get the tilde "~" symbol to generate.

Section 4.6 also replaced much of it's image-embedded text with actual words and tables.  Section 4.6 needs help though.  The figure provided is not very helpful, and there are a few random definitions provided with no context (and some of them are repeated from the beginning of Section 4).